requirements.txt
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.0
pytz==2023.3
python-docx>=0.8.11